+++ tests/core/scalar/test_parameter.py
@@ -16,6 +16,8 @@ from wetterdienst.provider.eccc.observation.metadata.dataset import (
     EcccObservationParameter,
 )
 from wetterdienst.provider.eccc.observation.metadata.unit import EcccObservationUnit
+from wetterdienst.provider.noaa.ghcn import NoaaGhcnParameter
+from wetterdienst.provider.noaa.ghcn.unit import NoaaGhcnUnit
 
 ORIGIN_UNITS = [unit.value for unit in OriginUnit]
 SI_UNITS = [unit.value for unit in SIUnit]
@@ -31,11 +33,17 @@ SI_UNITS = [unit.value for unit in SIUnit]
         (DwdMosmixUnit, False),
         (EcccObservationParameter, False),
         (EcccObservationUnit, False),
+        (NoaaGhcnParameter, False),
+        (NoaaGhcnUnit, False),
     ),
 )
 def test_parameter_names(parameter_enum, is_ds_tree):
     """Test parameter and dataset tree enums for consistent parameter naming following the
     core Parameter enum. Due to equal structure units are also tested here"""
+
+    def _check_quality_flags(param):
+        return param.startswith("QUALITY") or param.startswith("QN")
+
     parameters = []
 
     for res in parameter_enum:
@@ -43,17 +51,13 @@ def test_parameter_names(parameter_enum, is_ds_tree):
             for dataset in parameter_enum[res]:
                 for parameter in parameter_enum[res][dataset]:
                     parameter_name = parameter.name
-                    if not parameter_name.startswith(
-                        "QUALITY"
-                    ) and not parameter_name.startswith("QN"):
+                    if not _check_quality_flags(parameter_name):
                         if parameter_name not in Parameter._member_names_:
                             parameters.append(parameter_name)
         else:
             for parameter in parameter_enum[res]:
                 parameter_name = parameter.name
-                if not parameter_name.startswith(
-                    "QUALITY"
-                ) and not parameter_name.startswith("QN"):
+                if not _check_quality_flags(parameter_name):
                     if parameter_name not in Parameter._member_names_:
                         parameters.append(parameter_name)